prov/psm3: fix issue when running gpudirect on gpu with small bar size
PSM3_RV_GPU_CACHE_SIZE rejects values > 4095 (eg. 4 GB)
Fix math to avoid 32 bit overflow

Properly rotate EP used for eviction and limit rotation to same user
opened EP to avoid races.

Use the proper value for is_buf_gpu_mem in out of order message handler.

Adjust return value from evict functions to be int64_t so can safely
return bytes evicted without overflow.

Use the master EP in the given opened EP for GdrCopy. This will be
within the proper progress_lock (avoiding races and no need for
additional locks) and will at least consolidate GdrCopy pinning to the
1st rail/QP of a given application EP.  This also has the benefit that
the OpenMPI 4.1 BTL will not be used for for any GPU pinning nor
registration.

Use an evict_some strategy which starts with the master rail of the
opened EP, also rotates among the rails/QPs in that EP (if not enough
evicted from master rail) and the if more needs to be evicted will
rotate among the all the other EPs while getting the proper locks. No
more than 3 evict calls per evict_some, but ultimately all EPs will be
visited if there is BAR space pressure.

Added workaround nvidia IOMMU corruption after failed pin pages in some
situations. In these cases the next successful pin returns the start of
GPU BAR Region 1 as it's physical address. So we gather these addresses
for each GPU at job start and then treat the return of this address by
any GPU Direct RV pinning operation as a fatal job failure so we prevent
further data corruption.

Add PSM3_GPU_PIN_CHECK as a chicken switch for pin BAR address checking
when disabled, errors reading GPU BAR space are also non-fatal. Also
obtain GPU BAR space size and check it against GPU cache size warn when
cache size >= BAR space size-32M

Improve wording of cache size warning and fix data size of max
recommended GPU cache size.

(cherry picked from commit 3125daccf468771c3cef24afef02af9bb3adfa37)

+++ prov/psm3/psm3/psm_verbs_mr.h
@@ -154,7 +154,7 @@ extern psm2_mr_cache_t psm2_verbs_alloc_mr_cache(psm2_ep_t ep,
 extern int psm2_verbs_mr_cache_allows_user_mr(psm2_mr_cache_t cache);
 
 #ifdef PSM_CUDA
-extern int psm2_verbs_evict_some(psm2_ep_t ep, uint64_t length, int access);
+extern int64_t psm2_verbs_evict_some(psm2_ep_t ep, uint64_t length, int access);
 #endif
 
 // pd can be the verbs_ep.pd or NULL to use the RV module's kernel pd
